wire up in accessor
